% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_mts.R
\name{assign_mts}
\alias{assign_mts}
\title{Assign HCP's to a MTS Component Group}
\usage{
assign_mts(master_list, data)
}
\arguments{
\item{master_list}{A n x 2 .xlsx file with n component groups and a list of
provider specialties, clinician titles, and provider types}

\item{data}{data frame which includes as fields:
\itemize{
\item ACCESS_USER_PROV_SPECIALTY
\item ACCESS_USER_CLINICIAN_TITLE
\item ACCESS_USER_PROV_TYPE
}}
}
\value{
A list with two elemments. The first is data.frame object/lookup
table consisting of the patient id, access user's id, the access user's role
type, title and specialty, and their final MTS component group assignment.
The second is a data.frame with the same variables as the lookup table, but
the combinations are all lacking assignments.
}
\description{
Take Level 1 MTS Master List, clean it, and apply it to access log data
to perform a hierarchical assignment algorithm.
}
