% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_patient_data.R
\name{clean_patient_data}
\alias{clean_patient_data}
\title{Clean up patient data}
\usage{
clean_patient_data(patient_data)
}
\arguments{
\item{patient_data}{Collected patient_data_\if{html}{\out{<site>}} from smart_cancer_care.db}
}
\description{
Clean up cases in patient data with multiple rows.
}
\details{
\itemize{
\item Clean up dates
\item Clean up cancer site and stage variables
\item Change DATE_LAST_PATIENT_CONTACT_OR_DEA to dx of next cancer site, if applicable
\item Change stages to date intervals.
}
}
