% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_notes_regression.R
\name{clinical_notes_regression}
\alias{clinical_notes_regression}
\title{Clinical notes estimation}
\usage{
clinical_notes_regression(
  data,
  outcome_variable = c("ed", "hospitalization")[1],
  method = c("normal_approx", "adaptive_mcmc")[1],
  mcmc_arguments,
  epsilon = 1e-05,
  max_iteration = 100,
  phi_boundary_epsilon = 0.01
)
}
\arguments{
\item{data}{List of lists.  Each element of the outer list corresponds to
running feature_extraction() for a patient and specific time window.  It
should also have appended as an element of the list the outcomes named
ed and hospitalization}

\item{outcome_variable}{character. Which outcome should be modeled?  "ed" or "hospitalization"}

\item{method}{character. Either "normal_approx" or "adaptive_mcmc".}

\item{mcmc_arguments}{optional list.  If running adaptive mcmc, supply a list
of arguments that correspond to \code{\link[adaptMCMC]{MCMC}}.}

\item{epsilon}{relative tolerance for convergence assessment}

\item{phi_boundary_epsilon}{The optimization of phi will run from
(eps, 1 - eps).}

\item{max_iterations}{integer.  Maximum number of iterations to run}
}
\value{
Object of class "cnregfit".  Contains a list with the following:
\itemize{
\item \code{alpha}: Posterior mode of coefficients for purely patient features
\item \code{beta}: Posterior mode of coefficients for note importance features
\item \code{gamma}: Posterior mode of coefficients for MTS group reading importance features.
These are bounded between 0 and 1, with higher values indicating more importance.
\item \code{phi}: Posterior mode of penalty for not reading notes.  Bounded between
0 and 1, with higher values indicating a higher penalty.
\item \code{method}: Which method (normal approximation or MCMC) was used.
\item \code{covariance}: (If \code{method = normal_approx}) Posterior covariance estimate.  Note that for the
bounded variables, this covariance is on the logit scale.
\item \code{posterior_draws}: (if \code{method = "adaptive_mcmc"}) Object returned from
\code{\link[adaptMCMC]{MCMC}}.  Retrieve samples via \code{object$posterior_draws$samples}.
Also, you can feed \code{posterior_draws} into \code{\link[adaptMCMC]{MCMC.add.samples}}
to obtain more MCMC samples by continuning the MCMC sampler.
\item \code{error}
}
}
\description{
Estimate the effect of patient features, note importance, and reader
importance on the outcome of unplanned healthcare utilization.  Use
future::plan() to parallelize.
}
