% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bp_network_data.R
\name{create_bp_network_data}
\alias{create_bp_network_data}
\title{Create network data for a patient}
\usage{
create_bp_network_data(
  data_pat,
  db,
  site = c("UCDavis", "UCSD", "UCLA")[2],
  include_med_students = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{db}{SQLite connection.}

\item{site}{Which site should we pull data for?  Should be
a character, namely one of "UCDavis", "UCSD", or "UCLA".}

\item{include_med_students}{if FALSE, med students' events are not kept.}

\item{verbose}{logical.}

\item{pat_id}{character giving the patient id}

\item{only_earliest_read}{If TRUE, duplicated Views (same note, same viewer)
are removed and only the earliest is kept.}
}
\value{
A \code{tidygraph} object.

The \strong{node tibble} contains the following:
\itemize{
\item \code{name}: Character corresponding to the access logs and patient data.
\item \code{earliest_appearance}: dttm
\item \code{type}: Character, either \code{"note"} or \code{"hcp"}.
}

The \strong{edge tibble} contains the following:
\itemize{
\item \code{from}: Integer referring back to the node tibble.
\item \code{to}: Integer referring back to the node tibble.
\item \code{EVENT_ACTION}: Character, either \code{"Modify"} or \code{"View"}.
\item \code{date_time}: dttm
}
}
\description{
This function creates a bipartite network for HCPs and
clinical notes.
}
\details{
NOTES:
\itemize{
\item The curated access logs data are sufficiently small for
UCDavis at least that this could be done faster by
collecting all the access logs first.  However, for
memory considerations I'm still doing the filtering on
the patient ID on the SQLite database and then collecting.
\item If the first entry in the data is a View, set the author to be
the first Modify.  If no modify exists, use note author id.  The new
modify row is set to be one year prior to the patient's first
CANCER_DIAGNOSIS_DATE
\item For Non-UCSD sites, change
\itemize{
\item med student prov type for sites other than UCSD!
\item values to look for in this_notes_alogs$ACCESS_USER_PROV_TYPE
and ACCESS_USER_CLINICIAN_TITLE for med student
}
\item Multiple author situations are handled as follows:
\itemize{
\item A new author leads to a change in the note ID going forward.
\item The new note ID has multiple authors.  Previous authors
carry forward.
\item A new View is created with the secondary author as the
reader, and the original authors as the authors.
}
}
}
