% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hcp_hcp_network.R
\name{create_hcp_hcp_network}
\alias{create_hcp_hcp_network}
\title{Compute HCP-HCP network}
\usage{
create_hcp_hcp_network(
  bp_network,
  time_interval = c(mdy("1899-12-30"), mdy("1-1-2100"))
)
}
\arguments{
\item{bp_network}{tbl_graph object derived from create_bp_network_data()}

\item{time_interval}{vector of length 2 dttm}
}
\value{
A tidygraph object with edge weights reflecting the number
of clinical notes written by one HCP and read by another HCP.
}
\description{
Take bipartite network information and convert to HCP-HCP network,
respecting certain time constraints for the edges.
}
