% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mtsgroup_mtsgroup_network.R
\name{create_mtsgroup_mtsgroup_network}
\alias{create_mtsgroup_mtsgroup_network}
\title{Create MTS group-MTS group network}
\usage{
create_mtsgroup_mtsgroup_network(
  pat_id,
  bp_network,
  end_time,
  inpatient_prop_thresh = 0.75,
  db,
  site = c("UCDavis", "UCSD", "UCLA")[1]
)
}
\arguments{
\item{pat_id}{character giving the patient id}

\item{bp_network}{tbl_graph object derived from create_bp_network_data()}

\item{end_time}{dttm last date of network construction}

\item{inpatient_prop_thresh}{Numeric between 0 and 1.  What proportion of
access logs should be during an inpatient stay to consider the HCP on an
inpatient team?  Default is 0.75 based on looking at histograms for UCDavis.}

\item{db}{SQLite connection.}

\item{site}{Which site should we pull data for?  Should be
a character, namely one of "UCDavis", "UCSD", or "UCLA".}
}
\value{
A tidygraph object with edge weights reflecting the number
of clinical notes written by one HCP and read by another HCP.
}
\description{
Take the bipartite information and create a network connecting
the MTS groups.  Connections and absences are now persistent over time.
}
