% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{feature_extraction}
\alias{feature_extraction}
\title{Feature extraction for both-sides regression modeling}
\usage{
feature_extraction(
  pat_id,
  bp_network,
  time_before_dx = 30,
  end_time,
  inpatient_treatment = c("augment_mts", "covariate")[1],
  inpatient_prop_thresh = 0.75,
  db,
  site = c("UCDavis", "UCSD", "UCLA")[1]
)
}
\arguments{
\item{pat_id}{character giving the patient id}

\item{bp_network}{tbl_graph object derived from create_bp_network_data()}

\item{time_before_dx}{integer.  Number of days prior to date of diagnosis.}

\item{end_time}{dttm}

\item{inpatient_treatment}{Character.  Either (1) "augment_mts", in which case
all MTS groups will have an inpatient equivalent, effectively doubling the
number of MTS groups; or (2) "covariate", in which case inpatient will be
added as a note-importance covariate}

\item{inpatient_prop_thresh}{Numeric between 0 and 1.  What proportion of
access logs should be during an inpatient stay to consider the HCP on an
inpatient team?  Default is 0.75 based on looking at histograms for UCDavis.}

\item{db}{SQLite connection.}

\item{site}{Which site should we pull data for?  Should be
a character, namely one of "UCDavis", "UCSD", or "UCLA".}
}
\description{
Find patient level variables, note level features, and group readership
indicators.
}
