% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cnregfit.R
\name{print.cnregfit}
\alias{print.cnregfit}
\title{Print method for class cnregfit}
\usage{
\method{print}{cnregfit}(x)
}
\arguments{
\item{x}{object of class cnregfit}
}
\description{
Print method for class cnregfit
}
