% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within_between.R
\name{within_between}
\alias{within_between}
\title{Compute within/between communication measures}
\usage{
within_between(
  id_pat,
  db,
  bp_path = "/Shared/ehr_networks/R01-SMART_cancer_care/patient_networks/ucsd/bipartite/",
  site = "UCSD",
  lookback = 4,
  showProgress = TRUE,
  date_start,
  date_end
)
}
\arguments{
\item{id_pat}{character PAT_OBFUS_ID}

\item{db}{smart_cancer_care.db connection}

\item{bp_path}{character path to objects from create_bp_network_data()}

\item{site}{character}

\item{lookback}{positive integer.  Number of weeks to look back prior to
study entry date}

\item{date_start}{optional.  lubridate object}

\item{date_end}{optional.  lubridate object}
}
\description{
Compute within/between communication measures
}
\details{
For each team at each time point, compute
\enumerate{
\item Team size: Number of team members
\item Information production: Number of notes written by team members
\item Within-team Information Reach: Number of note reads (excluding re-reading)
from notes written by other team members
\item Within-team Information share: Proportion of reads that are actualized
from notes written by someone else on the same team
5a. Missed opportunities: Number of potential note reads - actual note reads
5b. Missed opportunities: Number of completely unread notes
}

and for each other team
6. Between-team Information Reach: Number of note reads (excluding re-reading)
from notes written by someone on the source team
7. Between-team Information Share: Proportion of reads that are actualized
from notes written by someone else on the source team
8a. Missed opportunities: Number of potential note reads - actual note reads
8b. Missed opportunities: Number of completely unread notes
}
