% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bp_network_data.R
\name{create_bp_network_data}
\alias{create_bp_network_data}
\title{Create network data for a patient}
\usage{
create_bp_network_data(
  pat_id,
  db,
  site = c("UCDavis", "UCSD", "UCLA")[1],
  include_med_students = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{pat_id}{character giving the patient id}

\item{db}{SQLite connection.}

\item{site}{Which site should we pull data for?  Should be
a character, namely one of "UCDavis", "UCSD", or "UCLA".}

\item{verbose}{logical.}
}
\value{
A \code{tidygraph} object.

The \strong{node tibble} contains the following:
\itemize{
\item \code{name}: Character corresponding to the access logs and patient data.
\item \code{earliest_appearance}: dttm
\item \code{type}: Character, either \code{"note"} or \code{"hcp"}.
}

The \strong{edge tibble} contains the following:
\itemize{
\item \code{from}: Integer referring back to the node tibble.
\item \code{to}: Integer referring back to the node tibble.
\item \code{EVENT_ACTION}: Character, either \code{"Modify"} or \code{"View"}.
\item \code{date_time}: dttm
}
}
\description{
This function creates a bipartite network for HCPs and
clinical notes.
}
\details{
NOTES:
\itemize{
\item The curated access logs data are sufficiently small for
UCDavis at least that this could be done faster by
collecting all the access logs first.  However, for
memory considerations I'm still doing the filtering on
the patient ID on the SQLite database and then collecting.
\item I have filtered out telephone encounters
\item Multiple author situations are handled as follows:
\itemize{
\item A new author leads to a change in the note ID going forward.
\item The new note ID has multiple authors.  Previous authors
carry forward.
\item A new View is created with the secondary author as the
reader, and the original authors as the authors.
}
}
}
